First version
This is the 'draft' if you will. Lots doesn't work but it's mostly a proof of concept that I can make a page with tailwind that's properly laid out and looks good. I'm just working on a better solution for smaller screen sizes as the 'sm:' small size for phones doesn't respond amazingly well.

Any feedback is welcome!